Move abstract_object_sett out into its own files.
Pull across functions from value_set_abstract_object and
value_set_pointer_abstract_object
