+++ src/analyses/Makefile
@@ -33,29 +33,29 @@ SRC = ai.cpp \
       static_analysis.cpp \
       uncaught_exceptions_analysis.cpp \
       uninitialized_domain.cpp \
-      variable-sensitivity/abstract_object.cpp \
       variable-sensitivity/abstract_environment.cpp \
+      variable-sensitivity/abstract_object.cpp \
+      variable-sensitivity/abstract_object_set.cpp \
+      variable-sensitivity/abstract_pointer_object.cpp \
       variable-sensitivity/abstract_value_object.cpp \
       variable-sensitivity/constant_abstract_value.cpp \
       variable-sensitivity/constant_pointer_abstract_object.cpp \
       variable-sensitivity/context_abstract_object.cpp \
-      variable-sensitivity/write_location_context.cpp \
-      variable-sensitivity/pointer_abstract_object.cpp \
-      variable-sensitivity/variable_sensitivity_domain.cpp \
-      variable-sensitivity/variable_sensitivity_object_factory.cpp \
-      variable-sensitivity/full_struct_abstract_object.cpp \
-      variable-sensitivity/full_array_abstract_object.cpp \
-      variable-sensitivity/write_stack.cpp \
-      variable-sensitivity/write_stack_entry.cpp \
       variable-sensitivity/data_dependency_context.cpp \
-      variable-sensitivity/value_set_abstract_object.cpp \
-      variable-sensitivity/variable_sensitivity_dependence_graph.cpp \
+      variable-sensitivity/full_array_abstract_object.cpp \
+      variable-sensitivity/full_struct_abstract_object.cpp \
       variable-sensitivity/interval_abstract_value.cpp \
+      variable-sensitivity/three_way_merge_abstract_interpreter.cpp \
       variable-sensitivity/value_set_abstract_object.cpp \
       variable-sensitivity/value_set_abstract_value.cpp \
       variable-sensitivity/value_set_pointer_abstract_object.cpp \
-      variable-sensitivity/three_way_merge_abstract_interpreter.cpp \
       variable-sensitivity/variable_sensitivity_configuration.cpp \
+      variable-sensitivity/variable_sensitivity_dependence_graph.cpp \
+      variable-sensitivity/variable_sensitivity_domain.cpp \
+      variable-sensitivity/variable_sensitivity_object_factory.cpp \
+      variable-sensitivity/write_location_context.cpp \
+      variable-sensitivity/write_stack.cpp \
+      variable-sensitivity/write_stack_entry.cpp \
       # Empty last line
 
 INCLUDES= -I ..
